+++ src/types/index.ts
@@ -252,11 +252,7 @@ type TresComponents = {
 }
 
 declare module 'vue' {
-  export interface GlobalComponents extends TresComponents { }
-}
-declare module '@vue/runtime-core' {
-  interface GlobalComponents extends TresComponents { }
-}
-declare module '@vue/runtime-dom' {
-  interface GlobalComponents extends TresComponents {}
+  export interface GlobalComponents extends TresComponents {
+    primitive: DefineComponent<TresPrimitive>
+  }
 }