+++ vite.config.ts
@@ -6,14 +6,13 @@ import copy from 'rollup-plugin-copy'
 import { defineConfig } from 'vite'
 import banner from 'vite-plugin-banner'
 import dts from 'vite-plugin-dts'
-import Inspect from 'vite-plugin-inspect'
+/* import analyze from 'rollup-plugin-analyzer' */
 
 import { coverageConfigDefaults } from 'vitest/config'
 
 /* import analyze from 'rollup-plugin-analyzer'
  */ /* import { visualizer } from 'rollup-plugin-visualizer' */
 import { bold, gray, lightGreen, yellow } from 'kolorist'
-
 import { resolve } from 'pathe'
 
 import pkg from './package.json'
@@ -42,7 +41,7 @@ export default defineConfig({
         pkg.version
       }\n * (c) ${new Date().getFullYear()}\n * description: ${pkg.description}\n * author: ${pkg.author}\n */`,
     }),
-    Inspect(),
+    // Inspect(),
   ],
   test: {
     environment: 'jsdom',
@@ -73,7 +72,7 @@ export default defineConfig({
           targets: [{ src: 'src/types/tres-components.d.ts', dest: 'dist/types' }],
         }),
         /*   analyze(), */
-        /*    visualizer({
+        /* visualizer({
           open: true,
           gzipSize: true,
           brotliSize: true,